API: Fix credit points

+++ src/artemis_client/api.py
@@ -253,7 +253,7 @@ class Feedback(BaseEntity, total=False):
     text: str
     detailText: str
     reference: str
-    credits: int
+    credits: float
     type: FeedbackType
     result: "Result"
     positive: bool
